pyproject.toml
[project]
name = "inventory-fastapi"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "SQLAlchemy[asyncio]~=2.0",
    "asyncpg",
    "alembic",
    "pydantic-settings~=2.4",
]
[tool.ruff]
line-length = 100